Reorganize some functions/packages.
Moved several generic functions out of device.go and
into util package.

Moved opencl functions out of miner and into device.go.
